reset like branch
